+++ src/main/java/nextstep/auth/util/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package nextstep.auth;
+package nextstep.auth.util;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
